Goomba added.
I created an enemy script for goomba which uses "Linecast" for casting a ray/line in a direction, using this i was able to add logic to goomba - he detects when he is at the edge of a cliff or when he needs to turn around. the rest of the script is simply giving a force in a direction to goomba, playing an animation that i created previously and flipping the direction goomba is facing in when he hits a wall or a cliff.

also added a death animation for goomba and made him die when colliding with the player, giving the player +100 score and adding a force.

created and added a run animation for mario.