stuff i did
Added immunity frames to mario, created a functioning menu where the user can select specific levels, click play and be taken to the first level, or exit the game. also did some work on the colliders and added knockback if the player touches an enemy, the player takes damage and gets immunity frames for 2 seconds.